scripts/committer_activities/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
datetime==5.4
    # via -r requirements.in
idna==3.6
    # via requests
mdutils==1.6.0
    # via -r requirements.in
numpy==1.26.4
    # via pandas
pandas==2.2.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   pandas
pytz==2024.1
    # via
    #   datetime
    #   pandas
requests==2.31.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via requests
zope-interface==6.2
    # via datetime

# The following packages are considered to be unsafe in a requirements file:
# setuptools
